Log rpc config on startup
